chore: move size to another file
